backend/requirements.txt
asgiref==3.6.0
attrs==22.2.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.7
django-cors-headers==3.14.0
django-import-export==3.1.0
django-nested-inline==0.4.6
djangorestframework==3.14.0
drf-extra-fields==3.4.1
drf-spectacular==0.26.1
drf-writable-nested==0.7.0
et-xmlfile==1.1.0
inflection==0.5.1
jsonschema==4.17.3
MarkupPy==1.14
numpy==1.24.2
odfpy==1.4.1
openpyxl==3.1.2
pandas==1.5.3
psycopg2==2.9.5
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
six==1.16.0
sqlparse==0.4.3
tablib==3.4.0
tzdata==2023.2
uritemplate==4.1.1
xlrd==2.0.1
XlsxWriter==3.0.9
xlwt==1.3.0
